+++ src/com/intellij/idea/plugin/hybris/acl/file/actions/AclOpenQueryAction.kt
@@ -1,0 +1,57 @@
+/*
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.intellij.idea.plugin.hybris.acl.file.actions
+
+import com.intellij.idea.plugin.hybris.acl.file.AclFileType
+import com.intellij.idea.plugin.hybris.actions.CopyFileToHybrisConsoleUtils
+import com.intellij.idea.plugin.hybris.common.HybrisConstants
+import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils.message
+import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
+import com.intellij.openapi.actionSystem.AnAction
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.actionSystem.CommonDataKeys
+import com.intellij.openapi.fileEditor.FileDocumentManager
+import com.intellij.psi.SingleRootFileViewProvider
+
+class AclOpenQueryAction : AnAction() {
+
+    init {
+        with (templatePresentation) {
+            text = message("hybris.acl.actions.open_query")
+            description = message("hybris.acl.actions.open_query.description")
+            icon = HybrisIcons.Console.Actions.OPEN
+        }
+    }
+
+    override fun actionPerformed(e: AnActionEvent) {
+        val project = e.project ?: return
+        val query = e.getData(CommonDataKeys.VIRTUAL_FILE_ARRAY)
+            ?.firstOrNull()
+            ?.takeIf { it.fileType is AclFileType }
+            ?.takeUnless { SingleRootFileViewProvider.isTooLargeForIntelligence(it) }
+            ?.let { FileDocumentManager.getInstance().getDocument(it) }?.text
+            ?: return
+
+        CopyFileToHybrisConsoleUtils.copyQueryToConsole(
+            project,
+            HybrisConstants.CONSOLE_TITLE_IMPEX,
+            query
+        )
+    }
+
+}