+++ src/com/intellij/idea/plugin/hybris/debugger/ui/tree/render/ModelEnumerationChildrenRendererInfoProvider.kt
@@ -1,6 +1,6 @@
 /*
- * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -26,9 +26,9 @@ import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.system.type.meta.TSMetaModelAccess
 import com.intellij.idea.plugin.hybris.system.type.meta.model.TSGlobalMetaItem
 import com.intellij.idea.plugin.hybris.system.type.meta.model.TSMetaRelation
-import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.project.Project
 import com.intellij.psi.search.GlobalSearchScope
+import com.intellij.util.application
 
 object ModelEnumerationChildrenRendererInfoProvider {
 
@@ -38,7 +38,7 @@ object ModelEnumerationChildrenRendererInfoProvider {
         className: String,
         fireRenderersChanged: Boolean = false
     ) {
-        ApplicationManager.getApplication().runReadAction {
+        application.runReadAction {
             val debuggerUtils = DebuggerUtils.getInstance()
             val typeCode = className
                 .substringAfterLast(".")