+++ src/com/intellij/idea/plugin/hybris/project/compile/ProjectAfterCompilerTask.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -21,14 +21,14 @@ import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.common.root
 import com.intellij.idea.plugin.hybris.common.yExtensionName
 import com.intellij.idea.plugin.hybris.settings.components.ProjectSettingsComponent
-import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.compiler.CompileContext
 import com.intellij.openapi.compiler.CompileTask
 import com.intellij.openapi.compiler.CompilerManager
+import com.intellij.util.application
 
 class ProjectAfterCompilerTask : CompileTask {
 
-    override fun execute(context: CompileContext) = ApplicationManager.getApplication().runReadAction<Boolean> {
+    override fun execute(context: CompileContext) = application.runReadAction<Boolean> {
         val settings = ProjectSettingsComponent.getInstance(context.project)
         if (!settings.isHybrisProject()) return@runReadAction true
         if (!settings.state.generateCodeOnRebuild) return@runReadAction true