+++ src/com/intellij/idea/plugin/hybris/project/view/HybrisProjectView.kt
@@ -30,13 +30,13 @@ import com.intellij.idea.plugin.hybris.project.descriptors.ModuleDescriptorType
 import com.intellij.idea.plugin.hybris.project.services.HybrisProjectService
 import com.intellij.idea.plugin.hybris.settings.components.ApplicationSettingsComponent
 import com.intellij.idea.plugin.hybris.settings.components.ProjectSettingsComponent
-import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.ProjectRootManager
 import com.intellij.openapi.vfs.VfsUtil
 import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.ui.SimpleTextAttributes
+import com.intellij.util.application
 import com.intellij.util.asSafely
 import java.io.File
 
@@ -107,7 +107,7 @@ open class HybrisProjectView(val project: Project) : TreeStructureProvider, Dumb
         val otherNodes = mutableListOf<AbstractTreeNode<*>>()
         val treeNodes = mutableListOf<AbstractTreeNode<*>>()
 
-        val projectService = ApplicationManager.getApplication().getService(HybrisProjectService::class.java)
+        val projectService = application.getService(HybrisProjectService::class.java)
         val projectRootManager = ProjectRootManager.getInstance(project)
         for (child in children) {
             if (child is PsiDirectoryNode) {